refctor: graph and graph ui plugins
