+++ src/graph/mod.rs
@@ -1,0 +1,4 @@
+pub mod graph;
+mod graph_ui;
+
+pub(crate) use graph::GraphPlugin;
